configuração Eslint e test
