+++ githubapi-frontend/src/app/shared/listar-cards/listar-cards.component.spec.ts
@@ -1,5 +1,6 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { ListarCardsComponent } from './listar-cards.component';
 
@@ -9,10 +10,9 @@ describe('ListarCardsComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ListarCardsComponent ],
-      imports: [HttpClientModule]
-    })
-    .compileComponents();
+      declarations: [ListarCardsComponent],
+      imports: [HttpClientModule, RouterTestingModule],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(ListarCardsComponent);
     component = fixture.componentInstance;
