+++ githubapi-frontend/src/app/shared/page/home/home.component.spec.ts
@@ -1,5 +1,8 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { BrowserModule } from '@angular/platform-browser';
+import { RouterTestingModule } from '@angular/router/testing';
+import { ServiceService } from '../../service/service.service';
 import { SharedModule } from '../../shared.module';
 
 import { HomeComponent } from './home.component';
@@ -11,8 +14,7 @@ describe('HomeComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [HomeComponent],
-      imports: [HttpClientModule, SharedModule]
-
+      imports: [HttpClientModule],
     }).compileComponents();
 
     fixture = TestBed.createComponent(HomeComponent);
