+++ githubapi-frontend/src/app/shared/page/user-profile/user-profile.component.html
@@ -1,3 +1,3 @@
 <div class="col container">
-    <h1>{{ dataUser?.name }}</h1>
+    <app-user-description></app-user-description>
 </div>
