+++ githubapi-frontend/src/app/shared/page/user-profile/user-profile.component.spec.ts
@@ -1,5 +1,6 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { UserProfileComponent } from './user-profile.component';
 
@@ -10,17 +11,14 @@ describe('UserProfileComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [UserProfileComponent],
-      imports: [HttpClientModule]
-    })
-      .compileComponents();
+      imports: [HttpClientModule, RouterTestingModule],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(UserProfileComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
-
-  it('should create HomeComponent', () => {
+  it('should create UserProfileComponent', () => {
     expect(component).toBeTruthy();
   });
-
 });
