+++ githubapi-frontend/src/app/shared/user-description/user-description.component.html
@@ -1,6 +1,7 @@
 <div class="col container">
     <div>
-        <img src="" alt="" {{ user.}}>
+        <!-- <span>{{ dataUser?.name }}</span> -->
+        <h1>{{ dataUser?.name }}</h1>
     </div>
 
 </div>
