+++ githubapi-frontend/src/app/shared/user-description/user-description.component.spec.ts
@@ -1,4 +1,6 @@
+import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ActivatedRoute, RouterModule } from '@angular/router';
 
 import { UserDescriptionComponent } from './user-description.component';
 
@@ -8,15 +10,14 @@ describe('UserDescriptionComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ UserDescriptionComponent ]
-    })
-    .compileComponents();
+      declarations: [UserDescriptionComponent],
+      imports: [HttpClientModule, RouterModule.forRoot([])],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(UserDescriptionComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
-
   it('should create', () => {
     expect(component).toBeTruthy();
   });
